Big update from MLP and CNN experiments. Dropout appears to be breaking CNN.

+++ src/catdog/catdog/utils/image.py
@@ -1,3 +1,4 @@
+import torch
 import matplotlib.pyplot as plt
 import matplotlib.image as mpimg
 import matplotlib.patches as patches
@@ -16,7 +17,7 @@ def plot_image_bbox(img, category, xmin, ymin, xmax, ymax, ax=None):
     xmax *= img.shape[0]
     ymin *= img.shape[1]
     ymax *= img.shape[1]
-    ax.imshow(img)
+    ax.imshow(img.type(torch.uint8))
     ax.add_patch(
         patches.Rectangle((xmin, ymin), xmax - xmin, ymax - ymin, linewidth=1.5, edgecolor='r', facecolor='none'))
     ax.annotate(category, (xmax + 5, ymax + 5), color="r", size=15)
